refactor: moved all taskwarrior handlers to package tw
Seperated Taskwarrior functions from handle_tw to tw package
