+++ backend/utils/tw/add_task.go
@@ -1,0 +1,54 @@
+package tw
+
+import (
+	"fmt"
+	"os"
+	"os/exec"
+)
+
+// add task to the user's tw client
+func AddTaskToTaskwarrior(email, encryptionSecret, uuid, description, project, priority, dueDate string) error {
+	cmd := exec.Command("rm", "-rf", "/root/.task")
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("error deleting Taskwarrior data: %v", err)
+	}
+
+	tempDir, err := os.MkdirTemp("", "taskwarrior-"+email)
+	if err != nil {
+		return fmt.Errorf("failed to create temporary directory: %v", err)
+	}
+	defer os.RemoveAll(tempDir)
+
+	origin := os.Getenv("CONTAINER_ORIGIN")
+	if err := SetTaskwarriorConfig(tempDir, encryptionSecret, origin, uuid); err != nil {
+		return err
+	}
+
+	if err := SyncTaskwarrior(tempDir); err != nil {
+		return err
+	}
+
+	cmdArgs := []string{"add", description}
+	if project != "" {
+		cmdArgs = append(cmdArgs, "project:"+project)
+	}
+	if priority != "" {
+		cmdArgs = append(cmdArgs, "priority:"+priority)
+	}
+	if dueDate != "" {
+		cmdArgs = append(cmdArgs, "due:"+dueDate)
+	}
+
+	cmd = exec.Command("task", cmdArgs...)
+	cmd.Dir = tempDir
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("failed to add task: %v", err)
+	}
+
+	// Sync Taskwarrior again
+	if err := SyncTaskwarrior(tempDir); err != nil {
+		return err
+	}
+
+	return nil
+}
