+++ backend/utils/tw/edit_task.go
@@ -1,0 +1,43 @@
+package tw
+
+import (
+	"fmt"
+	"os"
+	"os/exec"
+	"strings"
+)
+
+func EditTaskInTaskwarrior(uuid, description, email, encryptionSecret, taskuuid string) error {
+	cmd := exec.Command("rm", "-rf", "/root/.task")
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("error deleting Taskwarrior data: %v", err)
+	}
+	tempDir, err := os.MkdirTemp("", "taskwarrior-"+email)
+	if err != nil {
+		return fmt.Errorf("failed to create temporary directory: %v", err)
+	}
+	defer os.RemoveAll(tempDir)
+
+	origin := os.Getenv("CONTAINER_ORIGIN")
+	if err := SetTaskwarriorConfig(tempDir, encryptionSecret, origin, uuid); err != nil {
+		return err
+	}
+
+	if err := SyncTaskwarrior(tempDir); err != nil {
+		return err
+	}
+
+	// Escape the double quotes in the description and format it
+	escapedDescription := fmt.Sprintf(`description:"%s"`, strings.ReplaceAll(description, `"`, `\"`))
+
+	cmd = exec.Command("task", taskuuid, "modify", escapedDescription)
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("failed to edit task: %v", err)
+	}
+
+	// Sync Taskwarrior again
+	if err := SyncTaskwarrior(tempDir); err != nil {
+		return err
+	}
+	return nil
+}
